+++ msd/msd-sim/src/Environment/ReferenceFrame.hpp
@@ -65,7 +65,7 @@ public:
    *
    * @param globalCoords 3xN matrix of coordinates in global frame (modified in place)
    */
-  void globalToLocalBatch(Eigen::Matrix3Xd& globalCoords) const;
+  void globalToLocalBatch(Eigen::Matrix3Xf& globalCoords) const;
 
   /**
    * @brief Transform a coordinate from this local frame to global frame
@@ -89,7 +89,7 @@ public:
    *
    * @param localCoords 3xN matrix of coordinates in local frame (modified in place)
    */
-  void localToGlobalBatch(Eigen::Matrix3Xd& localCoords) const;
+  void localToGlobalBatch(Eigen::Matrix3Xf& localCoords) const;
 
   /**
    * @brief Set the origin of this frame in global coordinates
@@ -124,7 +124,7 @@ public:
    * @brief Get the rotation matrix
    * @return Const reference to the 3x3 rotation matrix
    */
-  const Eigen::Matrix3d& getRotation() const { return rotation_; }
+  const Eigen::Matrix3f& getRotation() const { return rotation_; }
 
   /**
    * @brief Get the origin (const version)
@@ -141,7 +141,7 @@ private:
   Coordinate origin_;  ///< Origin of this frame in global coordinates
   EulerAngles euler_;  ///< Rotation matrix from global to local frame
 
-  Eigen::Matrix3d rotation_;
+  Eigen::Matrix3f rotation_;
 };
 
 }  // namespace msd_sim
